155 - Create quartz job for DataOutThread
Create quartz job for DataOutThread.

155 - Create quartz job for DataOutThread

Fixed the test.

Rebase the changes for create quartz job for DataOutThread.
