+++ patientview-parent/patientview/src/main/java/org/patientview/service/UKTransplantManager.java
@@ -43,4 +43,6 @@ public interface UKTransplantManager {
     void save(UktStatus uktStatus);
 
     void deleteAll();
+
+    void exportPatientData() throws Exception;
 }
