+++ patientview-parent/patientview/src/main/resources/patientview.properties
@@ -61,6 +61,8 @@ ukt.directory=${ukt.directory}
 uktexport.directory=${uktexport.directory}
 ukt.archive.directory=${ukt.archive.directory}
 uktexport.archive.directory=${uktexport.archive.directory}
+dataout.directory=${dataout.directory}
+dataout.directory.thirdpart=${dataout.directory.thirdpart}
 
 # Messaging backend jobs
 quartz.sendMail.cron=${quartz.sendMail.cron}
